DEPLOYMENT/requirements.txt
streamlit==1.53.0
tensorflow==2.20.0
numpy
pillow